[Acexy] Middlestream: Make timeout configuration adjustable
When the AceStream middleware cannot connect against the given
stream, it used to hang undefinetely as the timeout was set to
`0` - which means "no timeout".

If a stream is not reachable anymore, this setting caused the
whole proxy to hang forever. A default value of `1s` is introduced
to cause the proxy to fail fast if the stream is not reachable. The
setting is made adjustable from the outside nonetheless, as depending
on the network connection, this may vary and affect the stream
latencies.